Extract `show_actions_panel` to `utils.py`
