+++ core/commands/push.py
@@ -8,6 +8,7 @@ from ...common import util
 from ..ui_mixins.quick_panel import show_remote_panel, show_branch_panel
 from ..ui_mixins.input_panel import show_single_line_input_panel
 from GitSavvy.core.runtime import enqueue_on_worker
+from GitSavvy.core.utils import show_actions_panel, noop
 
 
 __all__ = (
@@ -202,35 +203,3 @@ class gs_push_to_branch_name(PushBase):
             force=self.force,
             force_with_lease=self.force_with_lease
         )
-
-
-MYPY = False
-if MYPY:
-    from typing import Callable, Sequence, NamedTuple, Tuple
-    Action = NamedTuple("Action", [("description", str), ("action", Callable[[], None])])
-    ActionType = Tuple[str, Callable[[], None]]
-
-else:
-    from collections import namedtuple
-    Action = namedtuple("Action", "description action")
-
-
-def show_actions_panel(window, actions):
-    # type: (sublime.Window, Sequence[ActionType]) -> None
-    def on_action_selection(idx):
-        # type: (int) -> None
-        if idx == -1:
-            return
-        description, action = actions[idx]
-        action()
-
-    window.show_quick_panel(
-        [action[0] for action in actions],
-        on_action_selection,
-        flags=sublime.MONOSPACE_FONT
-    )
-
-
-def noop(description):
-    # type: (str) -> Action
-    return Action(description, lambda: None)
